refactor to allow for feeder robot in future update
