+++ pylitterbot/__init__.py
@@ -2,6 +2,6 @@
 __version__ = "2022.7.1b"
 
 from .account import Account
-from .robot import LitterRobot3, LitterRobot4, Robot
+from .robot import LitterRobot, LitterRobot3, LitterRobot4, Robot
 
-__all__ = ["Account", "Robot", "LitterRobot3", "LitterRobot4"]
+__all__ = ["Account", "Robot", "LitterRobot", "LitterRobot3", "LitterRobot4"]
