+++ pylitterbot/robot/__init__.py
@@ -1,5 +1,6 @@
 """pylitterbot robot module"""
 
+from .litterrobot import LitterRobot
 from .litterrobot3 import DEFAULT_ENDPOINT, DEFAULT_ENDPOINT_KEY, LitterRobot3
 from .litterrobot4 import LR4_ENDPOINT, LitterRobot4
 from .models import LITTER_ROBOT_4_MODEL
